+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * `AppOptions` now implements `Default`
 * added `InputApi.keys_released()` and `InputApi.keys_pressed()` that return iterators on key events since last update
 * added text input support through `InputApi.text()` (see text_input example)
+* added alpha demo showcasing framebuffer overdrawing
 
 # 1.2.0 - 2019 Nov 22
 ## fixed